god is dead and git killed them
