+++ multisig-wallet/project/SPECIFICATION.md
@@ -5,12 +5,18 @@ Table of Contents
     - [`cancel_transaction()`](#cancel_transaction)
     - [`constructor()`](#constructor)
     - [`execute_transaction()`](#execute_transaction)
+    - [`set_threshold()`](#set_threshold)
+    - [`set_weight()`](#set_weight)
     - [`transfer()`](#transfer)
   - [State Checks](#state-checks)
-    - [`nonce()`](#nonce)
+    - [`approval_weight()`](#approval_weight)
     - [`balance()`](#balance)
+    - [`nonce()`](#nonce)
+    - [`threshold()`](#threshold)
   - [Utilities](#utilities)
     - [`transaction_hash()`](#transaction_hash)
+    - [`threshold_hash()`](#threshold_hash)
+    - [`weight_hash()`](#weight_hash)
   - [Sequence Diagram](#sequence-diagram)
 
 # Overview
@@ -36,12 +42,10 @@ If you are interested in a functional overview then this is the section for you.
 
 1. Sets the parameters for approving a transaction and sets the owners of the multisig.
    1. If the constructor hasn't already been called.
-   2. Requires the `threshold`; the number of approvals required for a transaction to occur.
-      1. If the `threshold` is not 0.
+   2. Requires the config time constant `THRESHOLD`; the number of approvals required for a transaction to occur.
+      1. If the `THRESHOLD` is not 0.
    3. Requires the `users`; the information about the owners of the multisig
-      1. If none of the owners have the 0th address (0x00000...).
-      2. If none of the owners are set to have an approval weighting (number of approvals per owner) of 0.
-      3. If the sum of the owners' approval weightings is a value larger than the `threshold` parameter. This prevents the contract being setup when the owners can never submit enough approvals to allow a transaction.
+      1. If the sum of the owners' approval weightings is a value larger than the `threshold` parameter. This prevents the contract being setup when the owners can never submit enough approvals to allow a transaction.
 
 ### `execute_transaction()`
 
@@ -55,6 +59,26 @@ If you are interested in a functional overview then this is the section for you.
    6. Requires `signatures`; The information for each of the signatures submitted to approve a specific transaction.
    7. Requires `to`; The recipient of the transaction to be executed.
    8. Requires `value`; The value sent in the transaction to be executed.
+
+### `set_threshold()`
+
+1. Changes the threshold required for execution of transactions.
+2. Reverts when:
+   1. The constructor has not been called.
+   2. The new threshold is zero.
+   3. The new threshold is greater than the total weight of the owners.
+   4. Signature recovery failed.
+   5. Recovered addresses are not in ascending order.
+   6. The number of approvals does not meet the threshold.
+
+### `set_weight()`
+
+1. Sets the weight of a user which may be able to vote on the execution of transactions.
+2. Reverts when:
+   1. The constructor has not been called.
+   2. Signature recovery failed.
+   3. The number of approvals does not meet the threshold.
+   4. The new total weighting is less than the threshold.
 
 ### `transfer()`
 
@@ -72,14 +96,22 @@ If you are interested in a functional overview then this is the section for you.
 
 ## State Checks
 
-### `nonce()`
+### `approval_weight()`
 
-1. Returns the current nonce of the contract.
+1. Returns the approval weight of a user.
 
 ### `balance()`
 
 1. Returns the contract's balance of the specified asset.
    1. Requires `asset_id`; The contract ID of the asset to check that balance of.
+
+### `nonce()`
+
+1. Returns the current nonce of the contract.
+
+### `threshold()`
+
+1. Returns the threshold for execution.
 
 ## Utilities
 
@@ -90,6 +122,14 @@ If you are interested in a functional overview then this is the section for you.
    2. Requires `nonce`; The nonce field of the transaction.
    3. Requires `to`; The recipient of the transaction.
    4. Requires `value`; The value sent in the transaction.
+
+### `threshold_hash()`
+
+1. Returns the hash of a transaction used to change the threshold for execution.
+
+### `weight_hash()`
+
+1. Returns the hash of a transaction used to change the weight of a user.
 
 ## Sequence Diagram
