+++ multisig-wallet/project/contracts/multisig-contract/Forc.toml
@@ -4,4 +4,7 @@ entry = "main.sw"
 license = "Apache-2.0"
 name = "multisig-contract"
 
+[constants]
+THRESHOLD = { type = "u64", value = "5" }
+
 [dependencies]
