+++ multisig-wallet/project/contracts/multisig-contract/src/data_structures/user.sw
@@ -1,0 +1,8 @@
+library user;
+
+pub struct User {
+    /// The wallet address of a user.
+    address: b256,
+    /// The number of approvals the user provides when approving.
+    weight: u64,
+}
