+++ multisig-wallet/project/contracts/multisig-contract/src/events.sw
@@ -1,5 +1,9 @@
 library events;
 
+dep data_structures/user;
+
+use user::User;
+
 pub struct CancelEvent {
     cancelled_nonce: u64,
     user: b256,
@@ -10,6 +14,15 @@ pub struct ExecutedEvent {
     nonce: u64,
     to: Identity,
     value: u64,
+}
+
+pub struct SetThresholdEvent {
+    previous_threshold: u64,
+    threshold: u64,
+}
+
+pub struct SetWeightEvent {
+    user: User,
 }
 
 pub struct TransferEvent {
