+++ multisig-wallet/project/contracts/multisig-contract/tests/functions/core/cancel_transaction.rs
@@ -1,20 +1,22 @@
 use crate::utils::{
-    interface::{cancel_transaction, constructor, nonce},
-    test_helpers::{default_users, setup_env, DEFAULT_THRESHOLD},
-    CancelEvent, VALID_SIGNER_PK,
+    interface::{
+        core::{cancel_transaction, constructor},
+        info::nonce,
+    },
+    setup::{default_users, setup_env, VALID_SIGNER_PK},
 };
 
 mod success {
 
-    use fuels::types::Bits256;
-
     use super::*;
+    use crate::utils::setup::CancelEvent;
+    use fuels::types::Bits256;
 
     #[tokio::test]
     async fn cancels_transaction() {
         let (_private_key, deployer, _non_owner) = setup_env(VALID_SIGNER_PK).await.unwrap();
 
-        constructor(&deployer.contract, default_users(), DEFAULT_THRESHOLD).await;
+        constructor(&deployer.contract, default_users()).await;
 
         let initial_nonce = nonce(&deployer.contract).await.value;
 
@@ -45,7 +47,7 @@ mod revert {
     async fn not_an_owner() {
         let (_private_key, deployer, non_owner) = setup_env(VALID_SIGNER_PK).await.unwrap();
 
-        constructor(&deployer.contract, default_users(), DEFAULT_THRESHOLD).await;
+        constructor(&deployer.contract, default_users()).await;
 
         cancel_transaction(&non_owner.contract).await;
     }
