+++ multisig-wallet/project/contracts/multisig-contract/tests/functions/core/mod.rs
@@ -1,0 +1,6 @@
+mod cancel_transaction;
+mod constructor;
+mod execute_transaction;
+mod set_threshold;
+mod set_weight;
+mod transfer;
