+++ multisig-wallet/project/contracts/multisig-contract/tests/functions/info/mod.rs
@@ -1,0 +1,7 @@
+mod approval_weight;
+mod balance;
+mod nonce;
+mod threshold;
+mod threshold_hash;
+mod transaction_hash;
+mod weight_hash;
